🙆 Load models using tinyobj (#9)
This makes it so the sample app loads objects through tinyobj and uses
the existing structures to push them to the GPU. Additionally, index
buffers have been changed to 32 bits to allow for models of significant
vertex counts.

+++ include/backend/VulkanDevice.h
@@ -72,7 +72,7 @@ class VulkanDevice
         return *m_VertexBuffers.emplace_back(std::make_unique<VertexBuffer>(bufferCreateInfo, m_Device, m_PhysicalDevice, commandBuffer));
     }
 
-    IndexBuffer &CreateIndexBuffer(std::vector<uint16_t> data);
+    IndexBuffer &CreateIndexBuffer(std::vector<uint32_t> data);
 
     template<typename T> 
     UniformBuffer &CreateUniformBuffer()
